BG animaton happens on new showAnimationCompletion block. Fixe for completion blocks not being nil'ed.
Still doesnt solve problem which is: On my iPhone4 device (not simulator) if the drop animation is dismissed before finished showing (and therefore queued), the hud pops back to the center before being removed by the completion block.  I've tried:

- disabling the CATransaction begin/end when it's queue (as the queue block is contained in the showAnimation's transaction)
- setting removedOnCompletion to NO in _dropAnimationOut
- hud.layer.removeAllAnimations before adding the dismiss animations
- breaking in the dismiss completion block.  The hud is already been moved back to centre so it's not the cleanupAfter method or anything happening in completion.
- Subtracting 0.15 from the BG fade animation duration (which == *Long for drop animation) which is the only thing that works but is a total hack and very timing sensitive.  I've commented this out in the committed code.

Would love to fix this but I need to move on with my project.  Can't figure out why this would only happen when queued.  Must be the showAnimation taking precedent somehow...

Side note: For the completion block(s), I've added code to nil them after they are called as was specified in the comments for dismissAnimationCompletion.